feat: add weather location to database
